types yo

+++ nlp_services/scripts/distributed_entities.py
@@ -20,7 +20,7 @@ def es_get(pageid):
 def main():
     args = get_args()
     all_pages = ListDocIdsService().get_value(args.wiki_id)
-    slice_size = len(all_pages)/args.percentage_pages
+    slice_size = int(len(all_pages)/args.percentage_pages)
     slice_start = range(0, len(all_pages), slice_size)[args.slice]
     my_pages = all_pages[slice_start:slice_start+slice_size]
     print "Working on", slice_size, "pages"
